requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.34
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
